pyproject.toml
[tool.poetry]
name = "auto-data-scientist"
version = "0.1.0"
description = "Industry-grade AI system for automated data science using LangGraph, LangChain, and Playwright"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"

# Shared dependencies
pydantic = "^2.0"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
pytest-cov = "^4.1"
black = "^23.0"
ruff = "^0.1"
mypy = "^1.5"

[tool.poetry.group.orchestrator.dependencies]
fastapi = "^0.104"
uvicorn = {extras = ["standard"], version = "^0.24"}
langgraph = "^0.0.20"
langchain = "^0.1"
langchain-core = "^0.1"
httpx = "^0.25"

[tool.poetry.group.browser_agent.dependencies]
playwright = "^1.40"
fastapi = "^0.104"
uvicorn = {extras = ["standard"], version = "^0.24"}

[tool.poetry.group.ml_worker.dependencies]
optuna = "^3.4"
scikit-learn = "^1.3"
pandas = "^2.1"
numpy = "^1.24"
fastapi = "^0.104"
uvicorn = {extras = ["standard"], version = "^0.24"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by black)
]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=libs --cov=services --cov-report=html --cov-report=term"

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
